src/main/java/com/hhls/controller/UserController.java
package com.hhls.controller;

import com.hhls.pojo.User;
import com.hhls.service.UserService;
import java.util.List;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseBody;

@Controller
public class UserController {

    @Autowired
    private UserService userService;

    @RequestMapping("/list")
    @ResponseBody
    public List<User> list() {
        List<User> list = userService.readAll();
        return list;
    }
}
